Create system to monitor spoilers project and auto download on intervals. Also, cleanup some misc warnings
add choose location for download

New time label WIP, better functionality to disable

Better naming and minutes>hours

Spoiler Season text file (we control) to enable/disable the timers, so they don't run rogue

cause spoiler download success to reload card database

+++ oracle/src/oraclewizard.cpp
@@ -316,8 +316,8 @@ void LoadSetsPage::actDownloadProgressSetsFile(qint64 received, qint64 total)
 {
     if(total > 0)
     {
-        progressBar->setMaximum(total);
-        progressBar->setValue(received);
+        progressBar->setMaximum(static_cast<int>(total));
+        progressBar->setValue(static_cast<int>(received));
     }
     progressLabel->setText(tr("Downloading (%1MB)").arg((int) received / (1024 * 1024)));
 }
